requirements-dev.txt
argh==0.26.2
autopep8==1.4.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dash==0.30.0
dash-auth==1.2.0
dash-core-components==0.38.0
dash-html-components==0.13.2
dash-renderer==0.15.0
decorator==4.3.0
Flask==1.0.2
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
idna==2.7
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
libsass==0.16.0
MarkupSafe==1.1.0
mccabe==0.6.1
nbformat==4.4.0
pandas==0.23.4
pathtools==0.1.2
plotly==3.4.1
pycodestyle==2.4.0
python-dotenv==0.9.1
pytz==2018.7
PyYAML==3.13
requests==2.20.1
retrying==1.3.3
six==1.11.0
squarify==0.3.0
traitlets==4.3.2
ua-parser==0.8.0
urllib3==1.24.1
watchdog==0.9.0
Werkzeug==0.14.1
wrapt==1.10.11
